Being proactive about token expiration
- Refactored all of the HttpClient code into a single class, instead of
a class, an HttpClient, and HttpClientExtensions. (Pulled portions of
this down from the 2.0 code)
- Made the lowest level API call check if the `AccessToken.IsExpired`
before making the request. This means that the token is now checked on
every request.
- Reduced the number of retries from 50 to 25 (which is still
super-excessive).

Note: In order for the token check to take place well before the token
actually expires (to prevent situations where the token is valid before
the request, but invalid by the time it gets to the server), this
checkin requires #1344 as well.
