+++ PokemonGoAPI/Enums/ApiOperation.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PokemonGoAPI.Enums
+{
+
+    public enum ApiOperation
+    {
+        Retry,
+        Abort
+    }
+
+}
