--- PokemonGoAPI/HttpClient/PokemonClient.cs
@@ -1,24 +1,0 @@
-﻿using System.Net;
-using System.Net.Http;
-using PokemonGo.RocketAPI.Helpers;
-
-namespace PokemonGo.RocketAPI.HttpClient
-{
-    public class PokemonHttpClient : System.Net.Http.HttpClient
-    {
-        private static readonly HttpClientHandler Handler = new HttpClientHandler
-        {
-            AutomaticDecompression = DecompressionMethods.GZip | DecompressionMethods.Deflate,
-            AllowAutoRedirect = false
-        };
-
-        public PokemonHttpClient() : base(new RetryHandler(Handler))
-        {
-            DefaultRequestHeaders.TryAddWithoutValidation("User-Agent", "Niantic App");
-            DefaultRequestHeaders.ExpectContinue = false;
-            DefaultRequestHeaders.TryAddWithoutValidation("Connection", "keep-alive");
-            DefaultRequestHeaders.TryAddWithoutValidation("Accept", "*/*");
-            DefaultRequestHeaders.TryAddWithoutValidation("Content-Type", "application/x-www-form-urlencoded");
-        }
-    }
-}
