+++ PokemonGoAPI/PokemonGoAPI.csproj
@@ -116,6 +116,7 @@
   <ItemGroup>
     <Compile Include="Client.cs" />
     <Compile Include="Constants.cs" />
+    <Compile Include="Enums\ApiOperation.cs" />
     <Compile Include="Enums\AuthType.cs" />
     <Compile Include="Enums\StatusCode.cs" />
     <Compile Include="Exceptions\AccessTokenExpiredException.cs" />
@@ -128,9 +129,7 @@
     <Compile Include="Exceptions\LoginFailedException.cs" />
     <Compile Include="Exceptions\PTCOfflineException.cs" />
     <Compile Include="Extensions\DateTimeExtensions.cs" />
-    <Compile Include="Extensions\HttpClientExtensions.cs" />
     <Compile Include="Helpers\Crypt.cs" />
-    <Compile Include="Helpers\HttpClientHelper.cs" />
     <Compile Include="Helpers\IDeviceInfos.cs" />
     <Compile Include="Helpers\JsonHelper.cs" />
     <Compile Include="Helpers\RandomHelper.cs" />
@@ -139,7 +138,7 @@
     <Compile Include="Helpers\S2Helper.cs" />
     <Compile Include="Helpers\Utils.cs" />
     <Compile Include="Helpers\Vector.cs" />
-    <Compile Include="HttpClient\PokemonClient.cs" />
+    <Compile Include="Interfaces\IApiFailureStrategy.cs" />
     <Compile Include="ISettings.cs" />
     <Compile Include="Logger.cs" />
     <Compile Include="Logging\ConsoleLogger.cs" />
